connector_importer: split source mixins to own file
