+++ connector_importer/models/sources/__init__.py
@@ -1,2 +1,3 @@
+from . import source_consumer_mixin
 from . import source_mixin
 from . import source_csv
